Uart printing sensor project added. easybits issue partially fixed so data is sent over openchirp

+++ LoRaBug_Firmware/Firmware_LoRaMAC/Services/bmxService.h
@@ -42,7 +42,7 @@ extern "C"
  */
 extern void bmxService_createTask(void);
 
-extern void getBmxData(uint32_t *data);
+extern void getBmxData(uint8_t *data);
 
 
 /*********************************************************************
